requirements.txt
pyopengl==3.1.5
pyopengl-accelerate==3.1.5
glfw==1.11.1
pyrr==0.10.3
pillow==7.1.2
numpy==1.18.4